fix: string hash portability across platforms with different char signedness (issue #882)
Hash::operation(const std::string_view&) tail loop indexed the string_view
directly, yielding plain char. On platforms where char is signed
(x86-64, macOS arm64, wasm32) a byte >= 0x80 sign-extended before the
shift; on platforms where char is unsigned (aarch64/riscv64 linux) it
did not, so string primary-key hashes differed between platforms and
'=' on non-ASCII keys silently returned no rows after moving a database
file across platforms.

Cast the byte through int8_t to sign-extend explicitly everywhere.
This matches the historical behavior on signed-char platforms, so
existing indexes on x86-64/macOS/wasm32 remain valid and unsigned-char
platforms now produce identical hashes. Keys whose byte length is a
multiple of 8 were already portable and are unaffected.

+++ src/include/function/hash/hash_functions.h
@@ -146,7 +146,11 @@ inline void Hash::operation(const std::string_view& key, common::hash_t& result)
     }
     uint64_t last = 0;
     for (size_t i = 0u; i < key.size() % 8; i++) {
-        last |= static_cast<uint64_t>(key[key.size() / 8 * 8 + i]) << i * 8;
+        // Sign-extend the byte explicitly so the hash is identical on platforms where plain
+        // `char` is signed (x86-64, macOS arm64, wasm32) and unsigned (aarch64/riscv64 linux).
+        // This matches the historical behavior on signed-char platforms, keeping existing
+        // index entries on those platforms valid.
+        last |= static_cast<uint64_t>(static_cast<int8_t>(key[key.size() / 8 * 8 + i])) << i * 8;
     }
     hashValue = lbug::function::combineHashScalar(hashValue, lbug::function::murmurhash64(last));
     result = hashValue;
